Turning Account-related fields into properties
We are turning all fields that are on the Account model into properties on the Owner.

This way, we can use the getter/setter flow to set those fields, and the proper methods can correctly set them on the Account model whenever they change or are used.

To do that, we will turn, for example:

```
    plan = models.TextField(null=True, default=FREE_PLAN_NAME)  # Really an ENUM in db
```

into

```
    _plan = models.TextField(
        db_column="plan", null=True, default=FREE_PLAN_NAME
    )  # Really an ENUM in db
```

Notice that all that happened is that the property changed its name. And we added a `db_column` to point the newly-named property to the same database column as before.

Then we do

```
    @property
    def plan(self):
        return self._plan

    @plan.setter
    def plan(self, val):
        self._plan = val
```

so calls like `owner.plan` and `owner.plan = val` still make the proper column be affected. This way we risk less breaking changes

The only risk here is on the places where database queries that are made using the Django ORM use this column. I wasn't able to find any.

In a future change we will change the above methods to also change the account.plan field to match this value.

And in some point in the future we will build a command and a migration to "solidify" the change.

Notice that this change generates a Django migration. The thing is that Django doesn't understand that we renamed the attribute while keeping the database column the same. So when we naively use its value, it produces a migration with the following SQL:

```
ALTER TABLE "owners" RENAME COLUMN "plan" TO "_plan";
ALTER TABLE "owners" RENAME COLUMN "_plan" TO "plan";
````

So we manipulate the migration file a little and use a `SeparateDatabaseAndState` operation. What it does is essentially say "Django, act like you did operations A B and C, but in reality do operations D E and F.". We give it a `state_operations` list, which tells Django to consider those operations for everything that it concerns on "current state of migrations" (which is used for example on further `makemigration` calls), and a `database_operations`, which is the actual operations that produce the SQL that will be run on the database.

So we took the original naively generated list, put it inside `state_operations` (so Django thinks we made all the operations it told us to), and put an empty array on `database_operations`, so we don't actually do any database migrations.

The effectiveness is confirmed by the generated SQL:

```
(apienv) ➜  api git:(thiago/e/plan-fields-renaming) ✗ python manage.py sqlmigrate codecov_auth 0006
BEGIN;
--
-- Custom state/database change combination
--
COMMIT;
```

And the `makemigrations` command result:

```
(apienv) ➜  api git:(thiago/e/plan-fields-renaming) ✗ python manage.py makemigrations codecov_auth
No changes detected in app 'codecov_auth'
```